test: split lvs pool and stress
The stress tests log a lot which we don't really need to care about most
of the time, so let's disable all but error.
