+++ io-engine/tests/lvs_pool.rs
@@ -3,18 +3,17 @@ use io_engine::{
     bdev::crypto::{Cipher, EncryptionKey},
     bdev_api::bdev_create,
     core::{
-        logical_volume::LogicalVolume, MayastorCliArgs, NvmeCliArgs, NvmfShareProps, PoolCliArgs,
-        Protocol, Share, ToErrno, UnshareProps, UntypedBdev,
+        logical_volume::LogicalVolume, MayastorCliArgs, NvmeCliArgs, PoolCliArgs, Protocol, Share,
+        ToErrno, UnshareProps, UntypedBdev,
     },
     grpc::v1::pool::pool_to_proto,
     lvm::dm_setup::DmState,
-    lvs::{LvolSnapshotOps, Lvs, LvsError, LvsLvol, PropName, PropValue},
+    lvs::{Lvs, LvsError, LvsLvol, PropName, PropValue},
     pool_backend::{PoolArgs, PoolBackend, PoolOps, ReplicaArgs},
     subsys::NvmfSubsystem,
 };
-use io_engine_api::v1::{
-    pool::{ListPoolOptions, Pool, PoolAlert, PoolAlertStatus, PoolAlerts, PoolErrors, PoolState},
-    replica::ListReplicaOptions,
+use io_engine_api::v1::pool::{
+    Pool, PoolAlert, PoolAlertStatus, PoolAlerts, PoolErrors, PoolState,
 };
 use once_cell::sync::OnceCell;
 use std::{pin::Pin, str::FromStr};
@@ -1198,213 +1197,4 @@ async fn hot_detach_retach_(guard: TestHotRmGuard, share: bool, io: bool) {
     let nexus = io_engine::bdev::nexus::nexus_lookup_mut("nx").unwrap();
     nexus.destroy().await.unwrap();
     lvs.destroy().await.unwrap();
-}
-
-#[tokio::test]
-async fn lvol_list() {
-    let ms = ms();
-
-    let pool_size = "4GiB";
-    let repl_size = 4 * 1024 * 1024;
-    let replicas = 8000;
-
-    use io_engine::pool_backend::PoolMetadataArgs;
-    let pool_args = PoolArgs {
-        name: "tpool".into(),
-        disks: vec![format!("malloc:///m?size={pool_size}")],
-        backend: PoolBackend::Lvs,
-        md_args: Some(PoolMetadataArgs {
-            max_expansion: Some("300GiB".into()),
-        }),
-        ..Default::default()
-    };
-
-    ms.spawn(async move {
-        let lvs_pool = Lvs::create_or_import(pool_args).await.unwrap();
-
-        for i in 1..=replicas {
-            let name = format!("replica-{i}");
-            let opts = ReplicaArgs::new(name, repl_size)
-                .wipe_super(false)
-                .thin(true);
-            let _repl = lvs_pool.create_lvol_with_opts(opts).await.unwrap();
-        }
-    })
-    .await;
-
-    // this could vary depending on the system where we're running, but this is large enough that
-    // it should run on weaker systems as well as low enough to ensure we're testing the fix.
-    let max_dur = std::time::Duration::from_millis(300);
-
-    // 1. this list is "quicker" as there's no nvmf subsystems
-    let list_tm = std::time::Instant::now();
-    ms.spawn(async move {
-        for lvs_pool in Lvs::iter() {
-            println!();
-            for _ in 0..100 {
-                let mut count = 0;
-                for _lvol in lvs_pool.lvols() {
-                    count += 1;
-                }
-                assert_eq!(count, replicas);
-            }
-
-            for lvol in lvs_pool.lvols() {
-                let _replica: io_engine_api::v1::replica::Replica = lvol.into();
-            }
-        }
-    })
-    .await;
-    let no_uri_elapsed = list_tm.elapsed();
-    println!("Lvol List: {no_uri_elapsed:?}");
-    assert!(no_uri_elapsed <= max_dur, "Listing replicas took too long");
-
-    // 2. we share all replicas, so each replica now must search its subsystems
-    ms.spawn(async move {
-        for lvs_pool in Lvs::iter() {
-            for mut lvol in lvs_pool.lvols() {
-                use io_engine::replica_backend::ReplicaOps;
-                lvol.share_nvmf(NvmfShareProps::new()).await.unwrap();
-            }
-        }
-    })
-    .await;
-
-    // 3. we have to iter but also convert each lvol to a Replica so we can exercise the subsystem listing
-    let list_tm = std::time::Instant::now();
-    ms.spawn(async move {
-        let list_tm = std::time::Instant::now();
-        for lvs_pool in Lvs::iter() {
-            for lvol in lvs_pool.lvols() {
-                let _replica: io_engine_api::v1::replica::Replica = lvol.into();
-            }
-        }
-        println!("Lvol List time: {:?}", list_tm.elapsed());
-    })
-    .await;
-    let uri_elapsed = list_tm.elapsed();
-    println!("Lvol List: {uri_elapsed:?}");
-    assert!(uri_elapsed <= max_dur, "Listing replicas took too long");
-
-    use io_engine_api::v1::replica::ReplicaRpcClient;
-    let mut h = ReplicaRpcClient::connect("http://localhost:10124")
-        .await
-        .unwrap();
-
-    let list_tm = std::time::Instant::now();
-    h.list_replicas(ListReplicaOptions::default())
-        .await
-        .unwrap();
-    let grpc_elapsed = list_tm.elapsed();
-    println!("gRPC Lvol List: {grpc_elapsed:?}");
-    assert!(
-        // adds some extra buffer for gRPC
-        grpc_elapsed <= (max_dur + std::time::Duration::from_millis(100)),
-        "Listing replicas took too long"
-    );
-
-    use io_engine_api::v1::pool::PoolRpcClient;
-    let mut h = PoolRpcClient::connect("http://localhost:10124")
-        .await
-        .unwrap();
-
-    let list_tm = std::time::Instant::now();
-    h.list_pools(ListPoolOptions::default()).await.unwrap();
-    let grpc_elapsed = list_tm.elapsed();
-    println!("gRPC Lvs List: {grpc_elapsed:?}");
-    assert!(
-        // adds some extra buffer for gRPC
-        grpc_elapsed <= (max_dur + std::time::Duration::from_millis(100)),
-        "Listing pools took too long"
-    );
-
-    ms.spawn(async move {
-        for lvs_pool in Lvs::iter() {
-            lvs_pool.destroy().await.unwrap();
-        }
-    })
-    .await;
-}
-
-#[tokio::test]
-async fn lvol_snap_list() {
-    let ms = ms();
-
-    let pool_size = "4GiB";
-    let repl_size = 4 * 1024 * 1024;
-
-    use io_engine::pool_backend::PoolMetadataArgs;
-    let pool_args = PoolArgs {
-        name: "tpool".into(),
-        disks: vec![format!("malloc:///m?size={pool_size}")],
-        backend: PoolBackend::Lvs,
-        md_args: Some(PoolMetadataArgs {
-            max_expansion: Some("300GiB".into()),
-        }),
-        ..Default::default()
-    };
-
-    ms.spawn(async move {
-        let lvs_pool = Lvs::create_or_import(pool_args).await.unwrap();
-
-        for i in 1..=1024 {
-            let name = format!("replica-{i}");
-            let opts = ReplicaArgs::new(name, repl_size)
-                .wipe_super(false)
-                .thin(true);
-            let repl = lvs_pool.create_lvol_with_opts(opts).await.unwrap();
-            if i > 10 {
-                continue;
-            }
-            use io_engine::core::SnapshotParams;
-            for j in 1..=256 {
-                let snapshot = repl
-                    .create_snapshot(SnapshotParams {
-                        entity_id: Some(format!("e-{i}-{j}")),
-                        parent_id: Some(repl.uuid()),
-                        txn_id: Some(format!("txn-{i}-{j}")),
-                        snap_name: Some(format!("snap-{i}-{j}")),
-                        snapshot_uuid: Some(uuid::Uuid::new_v4().to_string()),
-                        create_time: Some(chrono::Utc::now().to_string()),
-                        discarded_snapshot: false,
-                    })
-                    .await
-                    .unwrap();
-                let n = uuid::Uuid::new_v4().to_string();
-                let _clone = snapshot
-                    .create_clone(io_engine::core::CloneParams {
-                        clone_name: Some(n.clone()),
-                        clone_uuid: Some(n.clone()),
-                        source_uuid: Some(snapshot.uuid()),
-                        clone_create_time: Some(chrono::Utc::now().to_string()),
-                    })
-                    .await
-                    .unwrap();
-            }
-        }
-    })
-    .await;
-
-    // this could vary depending on the system where we're running, but this is large enough that
-    // it should run on weaker systems as well as low enough to ensure we're testing the fix.
-    let max_dur = std::time::Duration::from_secs(3);
-
-    let list_tm = std::time::Instant::now();
-    ms.spawn(async move {
-        use io_engine::lvs::Lvol;
-        for snap in Lvol::list_all_snapshots(None) {
-            assert_eq!(snap.info().num_clones, 1);
-        }
-    })
-    .await;
-    let elapsed = list_tm.elapsed();
-    println!("Snapshot List: {elapsed:?}");
-    assert!(elapsed <= max_dur, "Listing snapshots took too long");
-
-    ms.spawn(async move {
-        for lvs_pool in Lvs::iter() {
-            lvs_pool.destroy().await.unwrap();
-        }
-    })
-    .await;
 }
